cid#398521 Initialization or destruction ordering is unspecified
and

cid#525184 Initialization or destruction ordering is unspecified
cid#525186 Initialization or destruction ordering is unspecified
