+++ common/Log.cpp
@@ -30,6 +30,7 @@
 #include <Poco/Version.h>
 
 #include "Log.hpp"
+#include "StaticLogHelper.hpp"
 #include "Util.hpp"
 
 namespace
@@ -370,73 +371,8 @@ namespace Log
         std::unordered_map<Poco::Message::Priority, std::string> _colorByPriority;
     };
 
-    /// Helper to avoid destruction ordering issues.
-    static struct StaticHelper
-    {
-    private:
-        GenericLogger* _logger;
-        static thread_local GenericLogger* _threadLocalLogger;
-        std::string _name;
-        std::string _logLevel;
-        std::string _id;
-        std::atomic<bool> _inited;
-    public:
-        StaticHelper() :
-            _logger(nullptr),
-            _inited(true)
-        {
-        }
-        ~StaticHelper()
-        {
-            _inited = false;
-        }
-
-        bool getInited() const { return _inited; }
-
-        void setId(const std::string& id) { _id = id; }
-
-        const std::string& getId() const { return _id; }
-
-        void setName(const std::string& name) { _name = name; }
-
-        const std::string& getName() const { return _name; }
-
-        void setLevel(const std::string& logLevel) { _logLevel = logLevel; }
-
-        const std::string& getLevel() const { return _logLevel; }
-
-        void setLogger(GenericLogger* logger) { _logger = logger; };
-
-        void setThreadLocalLogger(GenericLogger* logger)
-        {
-            // FIXME: What to do with the previous thread-local logger, if any? Will deleting it
-            // destroy also its channel? That won't be good as we use the same channel for all
-            // loggers. Best to just leak it?
-            _threadLocalLogger = logger;
-        }
-
-        GenericLogger* getLogger() const { return _logger; }
-
-        GenericLogger* getThreadLocalLogger() const { return _threadLocalLogger; }
-
-    } Static;
-
-    static struct StaticUIHelper: StaticHelper
-    {
-    private:
-        bool _mergeCmd = false;
-        bool _logTimeEndOfMergedCmd = false;
-    public:
-        void setLogMergeInfo(bool mergeCmd, bool logTimeEndOfMergedCmd)
-        {
-            _mergeCmd = mergeCmd;
-            _logTimeEndOfMergedCmd = logTimeEndOfMergedCmd;
-        }
-        bool getMergeCmd() const { return _mergeCmd; }
-        bool getLogTimeEndOfMergedCmd() const { return _logTimeEndOfMergedCmd; }
-    } StaticUILog;
-
-    thread_local GenericLogger* StaticHelper::_threadLocalLogger = nullptr;
+    extern StaticHelper Static;
+    extern StaticUIHelper StaticUILog;
 
     bool IsShutdown = false;
