+++ common/Util.hpp
@@ -40,6 +40,8 @@
 
 #include <StringVector.hpp>
 
+#include <config.h>
+
 #define STRINGIFY(X) #X
 
 #if CODE_COVERAGE
