+++ config.h.in
@@ -1,3 +1,5 @@
+#pragma once
+
 /* Variables that usually don't change once configured */
 
 /* The user-visible name of the app you build. */
