+++ net/clientnb.cpp
@@ -18,6 +18,8 @@
 #include <memory>
 #include <assert.h>
 
+#include <common/Globals.hpp>
+
 #include <Poco/Net/HTMLForm.h>
 #include <Poco/Net/HTTPClientSession.h>
 #include <Poco/Net/HTTPSClientSession.h>
