+++ tools/Benchmark.cpp
@@ -16,6 +16,7 @@
 
 #include <chrono>
 
+#include <common/Globals.hpp>
 #include <common/Png.hpp>
 #include <kit/Delta.hpp>
