+++ tools/KitClient.cpp
@@ -28,6 +28,7 @@
 #include <Poco/Util/Application.h>
 
 #include <KitHelper.hpp>
+#include <Globals.hpp>
 #include <Png.hpp>
 #include <Util.hpp>
 #include <Protocol.hpp>
