+++ wsd/COOLWSD.cpp
@@ -96,6 +96,7 @@
 #include <Poco/Util/XMLConfiguration.h>
 
 #include <common/Anonymizer.hpp>
+#include <common/Globals.hpp>
 #include <ClientRequestDispatcher.hpp>
 #include <Common.hpp>
 #include <Clipboard.hpp>
